+++ ydb/core/tx/columnshard/tablet/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    write_queue.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/hooks/abstract
+)
+
+END()
