+++ ydb/core/tx/columnshard/ya.make
@@ -30,7 +30,6 @@ GENERATE_ENUM_SERIALIZATION(columnshard.h)
 GENERATE_ENUM_SERIALIZATION(columnshard_impl.h)
 
 PEERDIR(
-    ydb/library/actors/core
     ydb/core/actorlib_impl
     ydb/core/base
     ydb/core/control
@@ -39,38 +38,40 @@ PEERDIR(
     ydb/core/protos
     ydb/core/tablet
     ydb/core/tablet_flat
-    ydb/core/tx/time_cast
-    ydb/core/tx/columnshard/engines
-    ydb/core/tx/columnshard/engines/writer
-    ydb/core/tx/columnshard/engines/reader/abstract
-    ydb/core/tx/columnshard/counters
-    ydb/core/tx/columnshard/common
-    ydb/core/tx/columnshard/splitter
-    ydb/core/tx/columnshard/operations
-    ydb/core/tx/columnshard/transactions
-    ydb/core/tx/columnshard/transactions/operators
-    ydb/core/tx/columnshard/blobs_reader
     ydb/core/tx/columnshard/blobs_action
+    ydb/core/tx/columnshard/blobs_action/storages_manager
+    ydb/core/tx/columnshard/blobs_reader
+    ydb/core/tx/columnshard/common
+    ydb/core/tx/columnshard/counters
+    ydb/core/tx/columnshard/data_accessor
+    ydb/core/tx/columnshard/data_accessor/in_mem
     ydb/core/tx/columnshard/data_locks
     ydb/core/tx/columnshard/data_sharing
-    ydb/core/tx/columnshard/subscriber
+    ydb/core/tx/columnshard/engines
+    ydb/core/tx/columnshard/engines/reader/abstract
+    ydb/core/tx/columnshard/engines/writer
     ydb/core/tx/columnshard/export
-    ydb/core/tx/columnshard/tx_reader
     ydb/core/tx/columnshard/loading
-    ydb/core/tx/columnshard/data_accessor
-    ydb/core/tx/columnshard/resource_subscriber
     ydb/core/tx/columnshard/normalizer
-    ydb/core/tx/columnshard/blobs_action/storages_manager
-    ydb/core/tx/columnshard/data_accessor/in_mem
-    ydb/core/tx/tiering
+    ydb/core/tx/columnshard/operations
+    ydb/core/tx/columnshard/resource_subscriber
+    ydb/core/tx/columnshard/splitter
+    ydb/core/tx/columnshard/subscriber
+    ydb/core/tx/columnshard/tablet
+    ydb/core/tx/columnshard/transactions
+    ydb/core/tx/columnshard/transactions/operators
+    ydb/core/tx/columnshard/tx_reader
     ydb/core/tx/conveyor/usage
-    ydb/core/tx/priorities/service
-    ydb/core/tx/tracing
     ydb/core/tx/long_tx_service/public
+    ydb/core/tx/priorities/service
+    ydb/core/tx/tiering
+    ydb/core/tx/time_cast
+    ydb/core/tx/tracing
     ydb/core/util
-    ydb/public/api/protos
-    ydb/library/yql/dq/actors/compute
+    ydb/library/actors/core
     ydb/library/chunks_limiter
+    ydb/library/yql/dq/actors/compute
+    ydb/public/api/protos
 )
 
 IF (OS_WINDOWS)
